requirements.txt
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
asyncpg==0.30.0
click==8.1.7
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.115.4
h11==0.14.0
httptools==0.6.4
idna==3.10
iso8601==2.1.0
jose==1.0.0
passlib==1.7.4
pyasn1==0.6.1
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pypika-tortoise==0.2.1
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2024.2
PyYAML==6.0.2
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
tortoise-orm==0.21.7
typing_extensions==4.12.2
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
